enable local server runs via stdio
