+++ src/servers/quickbooks/__init__.py
@@ -1,3 +1,4 @@
-from .main import create_server, get_initialization_options
+# Export server information without direct imports
+# This avoids circular imports when used through local.py
 
 __all__ = ['create_server', 'get_initialization_options']
